точк

+++ Para/Pizza.cs
@@ -8,7 +8,7 @@ namespace Para
 {
     internal static class Pizza
     {
-        public static int sum = 0;
+        public static double sum = 0;
         public static string pizza;
         public static string razmer;
         public static string dobavki;
